gml-server: Add a stub for parsing HTTP requests
Now whenever a connection is made a source for the socket is added to
the main context. When data is received it is given to an HTTP parser
instance that is specific to that connection. There are stub handlers
for the vtable functions of the parser that just g_print what
happened.