gml-server: Handle new person requests
The server now handles requests to GET /new_person by creating a
person in its person set and responding via GmlNewPersonResponse. Each
connection now has a queue of response objects which get queried for
data whenever the socket is ready for writing. If there is an HTTP
parse error then the queue is cleared and replaced with an error
response. Any further input is then ignored.

The server tries to support graceful shutdown by not closing the
connection until the client shuts down its end and initiating a
shutdown itself if it finishes sending an error response.