Ensure every GmlPerson is in a GmlConversation
The constructor for GmlPerson now takes a conversation so that every
person is always in a conversation. The server now expects there to be
a query string for the new_person request which it uses as the room
name. There is a hash table of pending conversations indexed by the
room name. The table only contains conversations that only have one
person. If another person is created with the same room name it will
grab the same conversation from the hash table.