gml-person: Add a 'use' count
This adds a count with a timer that is used to mark when the person is
in use. The idea is that when a connection is following a person the
use count will be incremented. Once the count reaches zero then the
timer is started. If the timer reaches 5 minutes then the person is a
candidate for removal.