Move the pending conversations hash table to a separate class
There's quite a lot of supporting code for the hash table of pending
conversations in the server which doesn't depend on any server state
so it makes sense for it to be a class of its own. That way we can
also get reference counting on it so that other objects can use it
without having to add API to the server object.