Move the request handlers to their own class
There is now a base class called GmlRequest which has virtual methods
for all of the vtable functions that GmlHttpParser has. The
request_finished method is expected to return a GmlResponse which the
server will then use. The two current methods have been adapted to be
subclasses of this class. This approach has the advantage that it
doesn't need to pick a response until after the complete request is
received and it can store state in between.