Responses can now delay sending data
This adds a new virtual to GmlResponse called 'has_data' which should
report whether the response has data immediately ready for sending. If
it returns FALSE then the request is waiting for more information
before any data will be sent. This will be used by the watch_person
handler to send messages only as they come in. GmlRespone now has a
signal to report when the has_data virtual may have changed. The
server uses this to decide whether to poll for writing. The default
implementation of has_data always returns TRUE.