Add the /watch_person request
This adds a request to watch for messages from a person. The messages
are sent via a new response which uses a chunked transfer encoding to
emit the messages as soon as they are available. The messages are
currently just concatenated together in text form but eventually it
needs to send them as JSON messages.