gml-person: Don't unref the conversation when the person leaves
Instead of unrefing the conversation and forgetting about it when the
person leaves the state of the conversation is now changed to mark
that it is over. This means the person can still download the
remaining messages. The conversation set can no longer use a weak
reference to notice when a conversation isn't waiting for a person but
it can instead listen to the changed signal and notice the state
change.