gml-leave-handler: Don't remove the person from the person set
Even though the conversation is over when the person leaves, it might
still be useful for the client to be able to watch the person to
notice that the conversation is over. This makes it so the only way a
person is destroyed is by the garbage collection that happens every 5
minutes.