Add an initial web site
This adds a web/ directory with the start of the web site. Currently
there is only the 'chat' page which jumps directly into a channel
called 'english'. There is some Javascript to perform a query and
listen for messages. The messages get added to the web page as they
come in.