gml-main-context: Report errors with a separate flag
Instead of conflating socket errors into GML_MAIN_CONTEXT_POLL_IN they
are now reported by a new GML_MAIN_CONTEXT_POLL_ERROR. When the server
sees this it now tries to get the error using getsockopt with SO_ERROR
but regardless of whether that succeeds it will always close the
connection. This is necessary because once an EOF is reported by the
read call returning 0 then Linux seems to continue to report 0 when
reading the socket even if there is a later error. This means that if
there is an error writing to the socket after a graceful shutdown has
been initiated then the server would get stuck in an infinite loop
because it would interpret the error condition as being ready for read
but it wouldn't get the error condition from the read call.