gml-conversation: Leave after 5 minutes of no messages
Instead of getting a person to leave the conversation after there has
been no use for 5 minutes, the GmlPersonSet now causes the
conversation to leave after there hasn't been a message for 5
minutes. This is necessary because the GmlWatchPerson response won't
notice that a connection is closed until it tries to write to the
socket or the TCP connection times out. If both people have actually
left the conversation and closed the socket then neither is going to
send a message so nothing will be written. The TCP connection can take
hours to timeout so the connection is effectively leaked.