Add support for translation
The files in the web directory are now filtered through a Perl script
which generates multiple versions of the file with different
replacements for each language. The original files now contain
placeholders for displayable strings that look like @THIS@. A separate
language file is parsed by the script which contains the replacement
values for each language.

The language files contain initial translations into English and
Esperanto.