gml-main-context: Add a default context
Instead of the server owning the context there is now a default
context which can be used by passing NULL wherever a main context is
expected. The default main context is now created in gml-main.c. To
make this a bit less fiddly all main context sources now store a
pointer back to their context and functions that operate on an
existing source don't need to specify the context.