gml-log: Separate opening the log file from starting the thread
If we want to be able to daemonize the server we ideally want to do
this after creating the log file in case we want to drop
privileges. However we don't want to launch a thread before forking so
we need to do that separately. This adds a gml_log_start function to
do just that.