Use a cached monotonic clock instead of GTimers
The GmlMainContext now has a function to receive the monotonic clock
time using g_get_monotonic_time. The value is cached so that we only
do the system call once between each call to epoll on the assumption
that the application doesn't block so it's as if no time passes
between polls.

GmlPerson and GmlConversation now uses this instead of a separate
GTimer allocation to track its staleness.