Drop connections that don't send a request within a timeout
The server now keeps track of the time a connection spends with no
response. If this value exceeds 5 minutes then it will send a timeout
response to the client and kill the connection. The timeout is only
checked every 5 minutes during the gc timer.

This is necessary to clean up clients that open a connection and then
don't send any data. If it didn't do this then they would just hang
around indefinetely and waste file descriptors.