gml-main-context: Don't use signalfd
signalfd is only available in glibc 2.8 which makes it awkward to use
in older distros. Instead GmlMainContext now just sets up a signal
handler the first time a quit callback is registered. The signal
handler puts a byte on a pipe to wake up the poll. This is needed
rather than just setting a flag to make it robust against the signal
being emitted between checking the flag and entering the poll.