Add methods to report when someone is typing
This adds handlers for /start_typing and /stop_typing. These update a
bitfield in the GmlConversation which stores which of the two
participants are typing. The GmlWatchPersonHandler sends an extra
command whenever the bit for the other person changes so that the
client knows whether he is typing. Sending a message explicitly marks
the person as no longer typing.