Assume text/plain in UTF-8 if there's no Content-Type header
When there is no Content-Type header in the /send_message request it
will now assume that the data will be text/plain in UTF-8. This is
necessary because when using XDomainRequest in Internet Exploiter it's
not possible to set the content type header or the charset. It does
however seem to send UTF-8 data anyway thankfully.