chat: Send a sneaky non-cross-domain sync request when using XDR
When trying to send a /leave request with XDomainRequest it's not
possible to do a synchronous request. To work around this it now also
sends an extra non-cross-domain synchronous request to the same URL as
the page using XMLHttpRequest. The hope is that this will cause enough
of a delay to flush the /leave request.