Use a request age for a person instead of a use count and a conv. age
Previously there was a timeout on the conversation which meant that it
would automatically be marked as finished if there is no message for
too long. Also a person would be removed if there is no connection
watching them for too long.

Both of these mechanisms have been replaced by a silence counter for
the person. If no request is received for a person after the timeout
then its conversation is ended and it is removed from the person set
so that it can't receive any more requests. Finishing the conversation
should also cause any watchers to finish so the person will likely be
immediately freed.

A separate empty request called 'keep_alive' has been added so that
clients can inform the server that they are still there. The clients
have been modified to send this if no message is sent after 2.5
minutes.

The advantage of this approach is that a user can wait for a partner
indefinitely and both users can keep a conversation going without
saying anything to each other if they feel like it.