Separate the interface language from the chat language
There are now separate links at the top of the page to choose the
interface language. The big buttons have been replaced with a drop
down box to choose from many more languages to chat in. The box
defaults to the language of the interface. ISO language codes are used
for the names of the rooms instead of the language name. The selected
language is passed to the next page in the search string via some
Javascript.