server: Switch to the Wayland model for signals
Instead of using GObject for signals, it now just uses the wayland
model. Each object that wants to expose a signal exposes a VsxSignal
struct which just contains a linked list of callbacks. The callbacks
are embedded list nodes which the callee uses as both a data pointer
and a marker to be able to later disconnect the listener.

This should make it simpler to add new signals and it should be faster
because it doesn't have all of the marshalling overhead of GObject.