server: Switch to a simpler object model
The server code doesn't need any of the features of GObject such as
properties and signals. The only thing it really uses is the
ref-counting, but that has a lot of overhead in gobject to implement
thread-safety which we don't need either. This patch makes it adopt a
simpler model. There is now a VsxObject struct which just contains a
reference count and a pointer to the vtable. The vtable for VsxObject
only contains a function to be called on destruction. The code can
make a subclass by greating a larger vtable struct that contains a
copy of the parent's vtable and initialising the object's vtable with
that.