watch-person-response: Split up the state for writing messages
The writing messages state was a bit confusing to deal with because if
there are no messages it would also handle updating the typing state.
However it wouldn't update the typing state unless it could fit the
entire message in the buffer because otherwise it could end up
starting to send a message in the middle of sending the typing state
update. This patch changes it so that there is are separate states for
waiting for new data, sending the typing state and sending messages.
Once one of these states is entered it won't be left until the data is
complete. That way the typing state doesn't have to have a complicated
check for whether there is space in the buffer. This approach also
makes it easy to send updates for other state changes such as the
details of a player.