Get rid of the public ‘awaiting start’ state
When there can be more than two people in the conversation it doesn't
really make much sense to prevent sending messages until someone
joins. In that case the start state doesn't really have much affect on
the public UI so we might as well drop it. There is still an internal
start state so that the server will know when to remove the
conversation from the list of pending conversations.

The ‘state’ message has been replaced with a simpler message with no
arguments called ‘end’ to mark that the conversation is over.