Accept more than two players in a conversation and send their state
The watch person response now sends state messages for each player
that include whether the player is connected and whether the player is
typing. The global ‘typing’ and ‘not-typing’ messages have been
removed. The conversation is no longer started when the second person
joins but instead it is now only started when the maximum number of
players in a room is reached. Eventually it should also be started the
first time a tile is flipped. The conversation finish state has been
removed and instead the players can leave a conversation but it does
not cause it to end. The conversation will naturally disappear when
all of the players have left and their connections have timed out.