web: Handle showing the player names on a board
The main page is now largely taken up with a blue board area which
will eventually hold the tiles. The board also contains 6 white
squares which represent the space reserved to store a player's tiles.
The name of the player is centered in this area. If there are more
than 6 players then their names are put in a separate div at the
bottom of the page.

The ChatSession object now keeps track of an object for each player
and stores its name, connectedness and typing state. When these
properties change it updates the CSS properties on the div for the
name. Currently this causes it to display a keyboard character when
the user is typing, and to strike a line through the name when the
player leaves the game.