server: Implement tile tracking
The conversation now keeps track of an array of tiles. These are
initialised with a fixed layout but the letters for the tiles are
shuffled.

The server now understands the requests ‘flip_tile’ and ‘move_tile’ to
update the tile state. These cause a signal to be emitted from the
conversation which the watch person response listens to to cause the
updated state to be sent.

The padding messages are no longer needed because the tile data is
always sent at the beginning of the response and this is around 9KB
which is enough to cause it to start invoking the onreadystatechange
handler.