web: Handle tile messages
The tile messages now cause a div to be created on the board to
represent the tile. It will either be empty or contain a letter
depending on the facing-up value. If a subsequent message is received
with a different position then it will use jQuery to animate the tile
to the new position.