web: Flip tiles when clicked
The web client now registers a handler for clicks in the board area.
When it detects that a click is on a tile it will queue ‘flip-tile’
message for that tile number. The message queue now contains array
pairs where the first part of the pair is the name of the command
instead of just always sending the send_message request.