web: Move flipped tiles that are clicked on next to last tile
In order to make it easier to lay tiles next to each other to make a
word, if you click on a tile that is already flipped it will move it
next to the last tile that was moved. That way you can move the first
letter of the word into place and then just click on the rest of the
letters.