web: Handle shouting
The UI now has a button next to the chat box for shouting. This will
send a shout request to the server. When a shout event is received the
shout button will be disabled for 10 seconds and the name of the
player who shouted will be shown in a large font centered in the
screen. That player's box will temporarily become red so that it is
obvious whose turn it is to make a word.