web: Try to automatically give focus to the input boxes
When the page is first loaded the input focus is now automatically set
to the name box. Once the game is connected it will also pass focus to
the chat box. It will also do this whenever the shout button is
pressed. The hope is that the chat box will generally retain focus so
typing a message should be quicker.