web: Throw an exception after setting the error state
Previously if we hit an error while processing messages the code would
carry on executing and restart the check data interval. That would end
up repeatedly creating errors when it tries to access this.watchAjax.
This patch just makes it throw one exception to stop executing
whatever it was doing. The page should then pretty much stop doing
anything.