server: Combine code for simple requests into a shared parent class
There are quite a few simple requests that just take a single argument
for the person ID and then call a single function on that person when
it is complete. These had some duplicated code to handle the argument
processing. This patch moves that code into a shared base class to
make it a bit easier to maintain.