server: Don't show unflipped tiles
When the game initially starts it will no longer report any tiles.
Instead of the /flip_tile request, there is now a /turn request which
doesn't take a tile number. This adds a new tile to the board which
will already be facing up. All available tiles are now always facing
up so the property has been removed from the tile message. The server
will find a location for the tile that does not overlap with any
existing tiles as close to the center as possible.