web: Add a button for turning tiles
This adds a button which will send the /turn request to the server. It
can also be activated by just pressing space when there is nothing in
the text entry box.

When a message for a new tile is found, instead of directly
positioning it and its new location, it is now animated in from the
top left corner of the screen.