server: Enforce the ordering for taking turns
The /turn requests are now ignored for players that don't have the
next turn or when any player is shouting.

The next turn flag is now more carefully moved to the next player when
a player leaves so that there will always be a player with the next
turn after the game has started.