web: Make the keep alive timeout more robust
I think that sometimes the keep alive callback was getting invoked
slightly early, which would mean the sendNextMessage function wouldn't
queue any message so no new keep alive would be queued.

This patch changes it so that the timeout always directly queues a
timeout message in the message queue so that it will definitely get
sent.

Instead of adding duplicated code for a third type of simple message,
this patch also makes it just take the message name directly as the
request method.