server: Detect first turn by number of tiles not state
The conversation can be marked as started even though there are no
tiles if the number of players reaches the maximum. In that case it
would previously not let anyone turn a tile because it would think
no-one had the next turn. This changes it to detect the first turn by
the number of tiles rather than the state to avoid that situation.