conversation: Keep track of the number of connected players
Previously when printing the log message when a game is started, it
would include the players that have left in the player count. This
fixes it to keep a separate count of the number of connected players.
This also fixes a bug where if some players leave a game before it
starts and it ends up with only one player, then that player won't
correctly receive the next turn when it starts.