web: Move the shout message to underneath the tiles and lower it a bit
It's kind of annoying that the shout message would cover up the tiles
making it difficult to see the letters that need to be dragged. This
just tweaks the z-indexes so the tiles will always appear on top and
moves the shout message down a bit so that it's usually below the
centre tiles.