web: Scale board and resize the messages so that it fills win height
Instead of the content being displayed at a constant height, the
JavaScript code will now try to make it exactly fill the height of the
window. To do this it will either scale down the board by setting the
font-size to a certain fraction of an em, or it will grow the message
box if there is too much space. The ChatSession now listens for
resizes events to recalculate the sizes.