web: Use the player property of “tile“ to ignore same user moves
Previously the web client would ignore tile position updates for the
tile that is currently being dragged. However sometimes if there is
some lag between the client and server then the messages would come
after the player has finished dragging. This would cause it to jump to
a previous position and then start animating to the place where the
tile was finally dropped. To fix that the web client now uses the new
‘player’ property to always ignore position updates for tiles that
were last moved by this player.