server: Add API to set the number of tiles used in the game
There is now a new request called /set_n_tiles which can be called
before the game has started to reduce the number of tiles that the
game will use. This is reported back to the client as a ‘n-tiles’
message. The intention is to give the players a chance to elect a
shorter game.