web: Add a drop-down box to select the game length
The start note now has a drop-down box to select the length of the
game which can either be a short game (50 tiles) or a normal game (122
tiles). Changing the selection immediately sends a /set_n_tiles
request which gets reflected in the UI for the other playes via the
n-tiles message. The server supports an arbritrary number of tiles so
the client picks which ever length of game is nearest for the
drop-down box.

The start note was previously a child of the board element which meant
that it was not possible to interact with it. To fix this there is now
a separate ‘scaled-area’ element which contains the start note so that
it will be scaled but does not prevent it from being interacted with.