server: Add a timeout mechanism to the main context
The main context can now have timeouts that get invoked at regular
intervals. The timeout value must be a multiple of a minute. To keep
the code simple, the timeouts that all have the same interval will be
grouped into a single bucket. That means that if the timeout is added
within the minute the first callback will be invoked in less than a
minute. The current intended use cases are just for things that should
happen vaguely regularly so it doesn't matter if it is not precise.

The garbage collection timer now uses this instead of manually
calculating the time to pass to vsx_main_context_poll().