web: Reconnect if no data is received after 2.5 minutes
The server now sends keep alive messages approximately every minute so
we can assume the connection is dead in some way if we don't get any
data for a while. I have a feeling this was causing some issues where
if a user waits for a long time for another player the connection can
be abandoned in some way by the network and the client will never
notice. If it detects a dead connection it will just silently
reconnect and carry on.