server: Don't add the GC source when there are no connections
The GC source has been split into two, one that is still owned by the
server and reaps connections with no request, and the other which is
owned by the VsxPersonSet which cleans up silent people. Both of these
are now removed if there are no connections or no people respectively
so that when nothing is happening the server can block without a
timeout.