server: Do a read when polling for read results in EPOLLHUP
It looks like if the server closes its end of the socket for writing
and continues to poll for read then when the client closes its end the
server will receive both EPOLLHUP and EPOLLIN. Previously this was
always being treated as an error but really it should be part of the
normal shutdown sequence. This change makes VsxMainContext report
EPOLLHUP as ready for reading when the client is requesting read so
that it has a chance to catch the HUP via a read which will return 0
and can gracefully close the socket instead of logging an error. If
EPOLLHUP is received in any other situation it will continue to be
considered an error.