server: Send a ‘sync’ message when all state has been sent
The server now sends a ‘sync’ message in response to a new_person or
watch_person request as soon as it has sent enough messages so that
the client is up-to-date with the current state of the player and the
server is going to wait for further state changes before sending any
more messages. The client can use this information to display messages
for further state changes such as when players enter or leave the
game.