web: Add a note in the message box when a player joins
When a new player joins or leaves the game there will now be a grey
message in the chat box to show at what point in the conversation they
joined. This will also end up making a sound and displaying a number
in the page title which makes it more likely that a waiting player
will notice that someone has joined.