web: Make the error messages more intrusive
Previously the error messages were fairly hidden at the top of the
screen so sometimes it would be hard to notice that the page is no
longer working while waiting for a game. This changes it to completely
obscure the page with a semi-transparent black rectangle and a large
error message. It no longer bothers to try and describe the detail of
the message because it's not particularly informative anyway. There is
now a button to reload the page after an error.

The ‘connecting’ status has been removed because it's not really
necessary and it looks out of place.