connection: Add dirty flags and remove the WRITE_WS_HEADER state
There’s no reason why the client couldn’t start sending frames directly
after sending the websocket negotiation so it’s better not to have a
separate state for it. Instead there is now a set of dirty flags for
things that the connection needs to send to the client and the WS header
is one of them.