main-context: Import the timer handling mechanism from notbit
Notbit has the extra ability that any timer source can be removed while
dispatching whereas Verda Ŝtelo was previously limited to removing only
the source that is currently being dispatched. VŜ doesn’t really need
the notbit mechanism but it also fixes a bug that occurs when removing
the currently dispatched source if it is the last one of its bucket.
Removing this last timer also frees the bucket and this would cause a
use-after-free error in the outer loop that iterates the buckets.