client/worker: Add a function to resolve the address
Moves the address resolving code from vsx_main to vsx_worker so that it
can be done on the worker thread. If the resolve fails it will retry
every 10 seconds. That way if the internet connection is temporarily
unavailable when the app is started it won’t just give up entirely.