client/connection: Calculate the wakeup timestamp in update_poll
Previously, update_poll only cached and recalculated the poll flags. If
anything else changes it was expected that the code would call
send_poll_changed directly. To make this a bit easier to handle,
update_poll now caches the entire poll_changed event and recalculates
everything. If anything is different then it will resend the event.

The main reason to do this is to make it possible to change the wakeup
time while handling events.