client/paint-state: Calculate the layout lazily
Instead of vsx_main calculating the layout just before painting, it is
now marked as dirty in the actual paint state sturct. There is now a
paint_state function to update the layout. That way the painters
themselves can trigger the layout update. This will be useful when the
layout is also used for handling input.