client/game-state: Add a signal for when the game state is modified
The signal is always emitted on the main thread. It is queued up for
emission whenever the game state receives an event which would need a
redraw. The idea is to handle the event processing for redraws in this
single place instead of having to do it in both the SDL and Android
frontends.