client/game-state: Handle events in the main thread
The connection events are now copied into a queue which is handled in
the main thread instead of trying to resync with the state from the
connection. This will make it possible to have a signal on the main
thread for every event and also avoid the need to have a copy of the
state in the vsx_connection.