client/game-state: Store the tile events instead of struct
In later patches, iterating the game state’s tiles is only going to be
used to get the tile painter up to date. The rest of the updates will be
handled by processing tile events. In order to make this a bit simpler,
the game state will now store the actual events used to update the tiles
so that the tile painter can reuse the same code to handle the
synchronisation as the updates.