client/tile-painter: Raise the drag tile instead of drawing it apart
Instead of having special code to ensure the dragged tile is always
above the other tiles, the tile painter now just moves it to the end of
the tile list whenever it is moved. This fixes a problem where the tile
painter doesn’t really know when the drag has finished and it was
previously keeping the dragged tile above tiles that have been moved by
someone else.