client/jni: Delay creating the worker
Creating the connection, worker and game state is now delayed to the
renderer thread. Otherwise this goes wrong when the worker tries to
queue a main thread idle event before the renderer constructor is
finished. This ends up trying to queue a command on the renderer thread
before the renderer is set on the surface and the app immediately
crashes.