server/conversation: Pass the tile data to the constructor
Instead of passing the room name in order to find the tile data for the
corresponding language, the conversation constructor now just directly
takes a pointer to the tile data. The conversation set handles finding
the tile data instead. This will make it easier to have conversations
with a different way of finding the tile data.