server/conversation: Add an ID to every conversation
The conversation set now has a hash table for the conversations so that
they can be found by ID, similar to the person IDs. The conversation set
now listens to a game until everyone has left instead of just until the
game starts. That way it will later be possible to join any game using
its ID, even if it has already started.