server/conversation-set: Put the pending conversations in separate list
The conversations that can be joined by room name are now kept in a
separate list that will hopefully remain relatively short. This is
important because the list is iterated to find a room with the same
name. It could be potentially quite slow to have to redundantly iterate
over all the conversations that are already in progress if there are a
lot of them.