client/connection: Allow room and player_name to be set later
The room and player_name are necessary if the connection is given a
person ID. Later patches will also add other situations like this.
Instead of setting these properties in the constructor, they are now set
with setter functions. Every time one of these is set the connection
will check if has enough configuration in order to start connecting.
Otherwise it will inform the poll to block indefinetely. This is an
expansion of the mechanism to wait for an address to connect to.