client/qr: Fix searching for bad patterns
The bad pattern is supposed to be a ratio so we also have to search for
doubled, tripled and quadrupled forms of the pattern. We don’t need to
search further than that because that would go over the size of the QR
code. It’s not totally clear from the spec but it probably also needs to
take into account the quiet zone around the code when checking for white
space either side of the pattern.