app: Store the vsx_instance_state in the bundle
Instead of directly storing the person ID in the bundle, the app now
stores the string returned by vsx_instance_state_save in order to make
it easier to later add more state. The state is protected by a mutex so
hopefully it should be ok to retrieve the string directly from the
Android UI main thread. Setting the state is queued to occur on the
rendering thread just like for the person ID.