client/game-state: Replace shouting flag with shouting player property
There can only be one player shouting, so instead of having a flag on
each player to mark whether they are shouting there is now just a
property on the game state to return the number of the player that is
currently shouting. This is easier to implement and just as easy to work
with.