client/shader-data: Store uniform locations for each program
Instead of having each painter have to manually retrieve the uniform
locations for the program it uses, the shader data itself now queries
all potential uniforms and stores the locations. It also handles
initialising the texture uniform to zero. This reduces a lot of code
duplication and should make it easier to have a painter that uses
multiple programs.