client/main-thread: Don’t store state in a static variable
Instead of having a single main thread in a state variable, it is now
created by the main layer and passed down to the game painter, the toolbox
and the image loader. This is mainly useful on Android where if an
intent is used to open another GameActivity then there will be a second
GameView with it’s own vsx_jni data in the same process. This will end
up using it’s own rendering thread. Previously this was causing havoc
when one of the activities is destroyed because it removed the wakeup
func for the other thread.