client/game-state: Handle the END message
The END message now just makes the game state reset like for the bad
ID errors except that it doesn’t set a note. The assumption is that this
should only happen as a result of the user requesting to leave the game
so they should be expecting the reset.