Add a script to get the freetype dependency
Instead of just assuming there is some freetype source code in a
freetype directory in the root of the source tree, there is now a script
to download a specific tarball. The tarball is checked against a known
sha256 hash. That way we can have the exact freetype version information
checked into the source tree.