client/game-state: Rename name_type to start_type
name_type was a weird name because the property has nothing to do with
the player name, it just happened that the only place that was using it
was the name dialog. Instead let’s name it so that it represents the
method used to start the game, ie, start a game or join another one.