client/tile-painter: Animate the clicked tiles immediately
A system of tile position overrides has been put in place with a timer
that reverts the positions back to what the server reports after 3
seconds. The dragging mechanism has been included into that too. That
way when a tile is clicked we can immediately give feedback to the user
that the click worked by moving the tile. If the server rejected it for
whatever reason then the tile will jump back in place after 3 seconds.