client/guide-painter: Don’t change width to fit the paragraph
Instead of changing the width of the dialog to fit the paragraph, we’ll
just keep the size we used to set the width of the layout. Otherwise
when the size changes the arrows at the sides of the dialog move around
and it’s annoying to browse the guide.