server/server: Send any pending responses after bad input
Instead of shutting down the write end of the connection immediately
after receiving bad input, the server now waits until the connection no
longer has any data to write. This gives it a chance to respond to any
messages that the client sent before it sent the bad input. This is
important if the client is trying to reconnect to a game that dosen’t
exist and doesn’t wait for the error message before also sending pending
tile updates. Previously that would have caused the client to miss the
error message and just keep reconnecting.