client/game-state: Add a function to reset and join game
Adds a version of vsx_game_reset that configures the connection to join
a conversation ID instead of reverting to the config to start a new
game. It will also set the appropriate start type. The plan is to use
this as the normal way of joining a game instead of having the upper
layers have to manually set the start type.