app: Use the singleTask launch mode
Otherwise weird things happen when you click on an invite link from
another application. The GameActivity would end up being opened in the
same task as the link. Then for example if you opened it in an IM
application you can’t get back to the chat without pressing back and
losing the game. This way you can switch between the chat and Verda
Ŝtelo using the task selector.