web: Implement invite URLs
The website can now handle invite URLs to join a game. The URLs contain
11 characters of base64-like data to contain the 64-bit conversation ID.
If the search string matches this pattern then it will be interpreted as
such instead of as a room ID. Currently there is no user interface to
display the invite link but it is printed on the console.