client/game-state: Set has_player_name when loading instance state
When the instance state sets a player ID on the connection then we can
assume the server has a name associated with the player. If we don’t do
this then after reloading the instance state, when a dialog is later
closed it will switch to showing the name dialog and mess up the game.