client/jni: Delete local references that aren’t returned to the VM
The local reference created from the weak reference to the surface is
created on a different thread that will never return to the VM. This was
causing the local reference table to overflow and crash the app when
testing on Android 5.1. The same thing could probably happen with the
share_link callback.