client/tile-painter: Clear the dragging tile if click on nothing
Previously if you click on the board where there isn’t a tile, it would
start dragging around whatever tile was last dragged because the
dragging_tile isn’t cleared. This fixes that to stop dragging whenever a
blank area of the board is clicked on and also if the drag start is
disabled because a player is shouting.